Rename module to random.
