+++ modules/random/CMakeLists.txt
@@ -3,7 +3,7 @@
 # =================================================================================================
 
 declare_module(
-  NAME testing
+  NAME random
   DEPENDS_ON_MODULES utils
   DEPENDS_ON_EXTERNAL_PROJECTS ""
 )
@@ -17,14 +17,14 @@ set(SOURCES
     src/random_generator.cpp
     src/random_type.cpp
     README.md
-    include/hephaestus/testing/random_container.h
-    include/hephaestus/testing/random_generator.h
-    include/hephaestus/testing/random_type.h
+    include/hephaestus/random/random_container.h
+    include/hephaestus/random/random_generator.h
+    include/hephaestus/random/random_type.h
 )
 
 # library target
 define_module_library(
-  NAME testing
+  NAME random
   PUBLIC_LINK_LIBS fmt::fmt magic_enum::magic_enum hephaestus::utils
   SOURCES ${SOURCES}
   PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>
